sandbox:updated sunburst visualizations
